Moving utils classes to verification package for clarity
