+++ integration-test-android/src/test/java/com/vimeo/integration_test_android/ComplexGenericClassTest.java
@@ -2,6 +2,8 @@ package com.vimeo.integration_test_android;
 
 import org.junit.Test;
 
+import verification.Utils;
+
 /**
  * Created by restainoa on 2/2/17.
  */
