+++ integration-test-android/src/test/java/com/vimeo/integration_test_android/UnserializablePlatformTypeTest.java
@@ -2,6 +2,8 @@ package com.vimeo.integration_test_android;
 
 import org.junit.Test;
 
+import verification.Utils;
+
 /**
  * Unit tests for {@link UnserializablePlatformType}.
  */
