+++ integration-test-android/src/test/java/verification/Utils.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.vimeo.integration_test_android;
+package verification;
 
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
