+++ integration-test-java-cross-module/src/test/java/verification/Utils.java
@@ -1,4 +1,4 @@
-package com.vimeo.sample;
+package verification;
 
 import android.support.annotation.NonNull;
