+++ integration-test-kotlin/src/test/kotlin/com/vimeo/dummy/sample_kotlin/KotlinGenericExampleTest.kt
@@ -1,6 +1,7 @@
 package com.vimeo.dummy.sample_kotlin
 
 import org.junit.Test
+import verification.Utils
 
 /**
  * Unit tests for [KotlinGenericExample]
